Make sky dark again.
This restores compatibility with shaders and  custom resource packs so that the sky actually looks like it's storming and raining when a storm is active. 

Personally I'd like to just see it as 1.0F so that things like Complementary shaders fully cover the built-in cloud layer for the immersion of it actually being a real storm. I don't think we'd see the sun with a F5 tornado on the ground or hailing super-cell :P

Even 0.8F would probably be fine. I understand the want for a less dramatic effect but it kind of just makes the mod feel more alive so to speak. Thanks ^_^